requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
tavily-python>=0.5.0
requests>=2.31.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
geopy>=2.4.0
langchain-community>=0.3.0
langgraph>=0.2.0
